+++ caseworker/advice/constants.py
@@ -30,5 +30,7 @@ class AdviceType:
 
 class AdviceSteps:
     RECOMMEND_APPROVAL = "recommend_approval"
+    DESNZ_APPROVAL = "desnz_approval"
+    FCDO_APPROVAL = "fcdo_approval"
     LICENCE_CONDITIONS = "licence_conditions"
     LICENCE_FOOTNOTES = "licence_footnotes"
