+++ caseworker/advice/urls.py
@@ -1,17 +1,19 @@
 from django.urls import path
 
 from caseworker.advice.views import views, consolidate_advice
+from caseworker.advice.views.add_advice import GiveApprovalAdviceView
+from caseworker.advice.views.edit_advice import EditAdviceView
 
 urlpatterns = [
     path("", views.AdviceView.as_view(), name="advice_view"),
     path("case-details/", views.CaseDetailView.as_view(), name="case_details"),
     path("select-advice/", views.SelectAdviceView.as_view(), name="select_advice"),
     path("approve-all-legacy/", views.GiveApprovalAdviceViewLegacy.as_view(), name="approve_all_legacy"),
-    path("approve-all/", views.GiveApprovalAdviceView.as_view(), name="approve_all"),
+    path("approve-all/", GiveApprovalAdviceView.as_view(), name="approve_all"),
     path("refuse-all/", views.RefusalAdviceView.as_view(), name="refuse_all"),
     path("view-my-advice/", views.AdviceDetailView.as_view(), name="view_my_advice"),
     path("edit-advice-legacy/", views.EditAdviceViewLegacy.as_view(), name="edit_advice_legacy"),
-    path("edit-advice/", views.EditAdviceView.as_view(), name="edit_advice"),
+    path("edit-advice/", EditAdviceView.as_view(), name="edit_advice"),
     path("delete-advice/", views.DeleteAdviceView.as_view(), name="delete_advice"),
     path("countersign/", views.CountersignAdviceView.as_view(), name="countersign_advice_view"),
     path("countersign/review-advice/", views.ReviewCountersignView.as_view(), name="countersign_review"),
